test: split tests by plugins
