+++ test/app.test.js
@@ -15,216 +15,30 @@
 'use strict';
 
 const assert = require('assert');
-const puppeteer = require('puppeteer');
 
-const debug = false;
-const toResp = (resp) => ({
-  status: resp ? 200 : 404,
-  // eslint-disable-next-line no-nested-ternary
-  body: resp ? (/^[{[]/.test(resp) ? JSON.stringify(resp) : resp) : '',
-});
-const apiMocks = {
-  dummy: {
-    webPath: '/dummy',
-  },
-  blog: {
-    webPath: '/en/topics/bla.html',
-    preview: {
-      lastModified: 'Fri, 18 Jun 2021 09:57:42 GMT',
-    },
-    source: {
-      lastModified: 'Fri, 18 Jun 2021 09:55:03 GMT',
-    },
-    edit: {
-      url: 'https://adobe.sharepoint.com/:w:/r/sites/TheBlog/_layouts/15/Doc.aspx?sourcedoc=%7BE8EC80CB-24C3-4B95-B082-C51FD8BC8760%7D&file=bla.docx&action=default&mobileredirect=true',
-    },
-  },
-  pages: {
-    webPath: '/creativecloud/en/test',
-    preview: {
-      lastModified: 'Fri, 18 Jun 2021 09:57:42 GMT',
-    },
-    source: {
-      lastModified: 'Fri, 18 Jun 2021 09:55:03 GMT',
-    },
-    edit: {
-      url: 'https://docs.google.com/document/d/2E1PNphAhTZAZrDjevM0BX7CZr7KjomuBO6xE1TUo9NU/edit',
-    },
-  },
-};
-const purgeMock = [{ status: 'ok' }];
-const htmlMock = '<html></html>';
+const {
+  IT_DEFAULT_TIMEOUT,
+  MOCKS,
+  getPlugins,
+  waitForEvent,
+  checkEventFired,
+  clickButton,
+  mockStandardResponses,
+  testPageRequests,
+  sleep,
+  getPage,
+  startBrowser,
+  stopBrowser,
+} = require('./utils');
+
+const fixturesPrefix = `file://${__dirname}/fixtures`;
 
 describe('Test sidekick bookmarklet', () => {
-  const IT_DEFAULT_TIMEOUT = 60000;
-  const fixturesPrefix = `file://${__dirname}/fixtures`;
-
-  const getPlugins = async (p) => p.evaluate(
-    () => Array.from(document.querySelectorAll('.hlx-sk > div.env > div, .hlx-sk > div:not(.env)'))
-      .map((plugin) => ({
-        id: plugin.className,
-        text: plugin.textContent,
-        buttonPressed: plugin.querySelector(':scope > button')
-          && plugin.querySelector(':scope > button').classList.contains('pressed'),
-      })),
-  );
-
-  const waitForEvent = async (p, type) => p.evaluate((evtType) => {
-    if (!evtType) return;
-    const evtTypes = typeof evtType === 'string' ? [evtType] : evtType;
-    evtTypes.forEach((et) => {
-      // set up test var and event listener
-      window[`${et}EventFired`] = false;
-      window.hlx.sidekick.addEventListener(et, () => {
-        window[`${et}EventFired`] = true;
-      });
-    });
-  }, type);
-
-  const checkEventFired = async (p, type) => p.evaluate(async (evtType) => {
-    if (!evtType) return true;
-    const evtTypes = typeof evtType === 'string' ? [evtType] : evtType;
-    const results = await Promise.all(evtTypes.map((et) => window[`${et}EventFired`]));
-    return results.every((res) => res === true);
-  }, type);
-
-  const execPlugin = async (p, id) => {
-    await waitForEvent(p, 'pluginused');
-    await p.evaluate((pluginId) => {
-      const click = (el) => {
-        const evt = document.createEvent('Events');
-        evt.initEvent('click', true, false);
-        el.dispatchEvent(evt);
-      };
-      if (pluginId) {
-        click(document.querySelector(`.hlx-sk .${pluginId} button`));
-      }
-    }, id);
-    assert.ok(await checkEventFired(p, 'pluginused'), 'Event pluginused not fired');
-  };
-
-  const clickButton = async (p, id) => p.evaluate((buttonId) => {
-    const click = (el) => {
-      const evt = window.document.createEvent('Events');
-      evt.initEvent('click', true, false);
-      el.dispatchEvent(evt);
-    };
-    click(window.document.querySelector(`.hlx-sk button.${buttonId}`));
-  }, id);
-
-  const mockStandardResponses = async (p, opts = {}) => {
-    const { pluginsJs = '', configJs = '', check = () => true } = opts;
-    await p.setRequestInterception(true);
-    p.on('request', (req) => {
-      if (req.url().endsWith('/tools/sidekick/plugins.js')
-        && check(req)) {
-        req.respond(toResp(pluginsJs));
-      } else if (req.url().endsWith('/tools/sidekick/config.js')
-        && check(req)) {
-        req.respond(toResp(configJs));
-      } else if (req.url().startsWith('https://admin.hlx3.page/')) {
-        req.respond(toResp(apiMocks.dummy));
-      } else {
-        // console.log(req.url());
-        req.continue();
-      }
-    });
-  };
-
-  const testPageRequests = async ({
-    page,
-    url,
-    plugin,
-    events,
-    prep = () => {},
-    check = () => true,
-    mockResponses = purgeMock,
-    checkCondition = (req) => req.url().startsWith('https://'),
-    timeout = 0,
-    timeoutSuccess = true,
-  }) => {
-    let count = 0;
-    await page.setRequestInterception(true);
-    return new Promise((resolve, reject) => {
-      // watch for new browser window
-      page.on('request', async (req) => {
-        if (req.url().endsWith('/tools/sidekick/plugins.js')
-          || req.url().endsWith('/tools/sidekick/config.js')) {
-          // send plugins response
-          return req.respond(toResp());
-        } else if (checkCondition(req)) {
-          try {
-            if (check(req)) {
-              // check successful
-              resolve();
-            }
-            if (req.url().startsWith('file://')) {
-              // let file:// requests through
-              // console.log(req.url());
-              req.continue();
-            } else {
-              // send mock response
-              // console.log(req.url());
-              const response = toResp(mockResponses[count] || '');
-              count += 1;
-              return req.respond(response);
-            }
-          } catch (e) {
-            reject(e);
-          }
-        }
-        // let request continue
-        // console.log(req.url());
-        return req.continue();
-      });
-      // open url and optionally click plugin button
-      page
-        .goto(url, { waitUntil: 'load' })
-        .then(() => prep(page))
-        .then(() => waitForEvent(page, events))
-        .then(() => execPlugin(page, plugin))
-        .then(() => checkEventFired(page, events))
-        .catch((e) => reject(e));
-      if (timeout) {
-        setTimeout(() => {
-          if (timeoutSuccess) {
-            resolve();
-          } else {
-            reject(new Error('check timed out'));
-          }
-        }, parseInt(timeout, 10));
-      }
-    });
-  };
-
-  const sleep = async (delay = 1000) => new Promise((resolve) => setTimeout(resolve, delay));
-
-  let browser;
-  let page;
-
-  beforeEach(async function setup() {
-    this.timeout(10000);
-    browser = await puppeteer.launch({
-      devtools: debug,
-      args: [
-        '--disable-popup-blocking',
-        '--disable-web-security',
-        '–no-sandbox',
-        '–disable-setuid-sandbox',
-      ],
-    });
-    page = await browser.newPage();
-  });
-
-  afterEach(async () => {
-    if (!debug) {
-      await browser.close();
-      browser = null;
-      page = null;
-    }
-  });
+  beforeEach(startBrowser);
+  afterEach(stopBrowser);
 
   it('Renders with missing config', async () => {
+    const page = getPage();
     await page.goto(`${fixturesPrefix}/config-none.html`, { waitUntil: 'load' });
     const skHandle = await page.$('div.hlx-sk');
     assert.ok(skHandle, 'Did not render without config');
@@ -238,6 +52,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Renders with irrelevant config', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/config-wrong.html`, { waitUntil: 'load' });
     const skHandle = await page.$('div.hlx-sk');
@@ -252,6 +67,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Checks for hlx3 config/URL mismatch', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await new Promise((resolve, reject) => {
       // wait for dialog
@@ -274,23 +90,8 @@ describe('Test sidekick bookmarklet', () => {
     });
   }).timeout(IT_DEFAULT_TIMEOUT);
 
-  it('Shows live plugin without host in hlx3 mode', async () => {
-    await mockStandardResponses(page);
-    // open test page and click preview button
-    await page.goto(`${fixturesPrefix}/config-hlx3-no-host.html`, { waitUntil: 'load' });
-    const plugins = await getPlugins(page);
-    // check for live plugin
-    assert.ok(plugins.find((plugin) => plugin.id === 'live'), 'Live plugin not shown');
-    // check outerHost
-    const outerHost = await page.evaluate(() => window.hlx.sidekick.config.outerHost);
-    assert.strictEqual(
-      outerHost,
-      'main--theblog--adobe.hlx.live',
-      `Outer CDN not defined as expected: ${outerHost}`,
-    );
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
   it('Uses main branch by default', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/config-no-ref.html`, { waitUntil: 'load' });
     const ref = await page.evaluate(() => window.hlx.sidekick.config.ref);
@@ -298,6 +99,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Adds plugin from config', async () => {
+    const page = getPage();
     await mockStandardResponses(page, {
       configJs: `
         window.hlx.initSidekick({
@@ -309,13 +111,16 @@ describe('Test sidekick bookmarklet', () => {
             }
           }],
         });`,
+      mockResponses: [MOCKS.api.blog],
     });
     await page.goto(`${fixturesPrefix}/config-default.html`, { waitUntil: 'load' });
+    await sleep();
     const plugins = await getPlugins(page);
     assert.ok(plugins.find((p) => p.id === 'foo'), 'Did not add plugin from config');
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Shows update dialog on legacy config', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await new Promise((resolve, reject) => {
       page.on('dialog', (dialog) => {
@@ -331,6 +136,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Shows update dialog in compatibility mode', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await new Promise((resolve, reject) => {
       page.on('dialog', (dialog) => {
@@ -347,6 +153,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Detects innerHost and outerHost from config', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/config-default.html`, { waitUntil: 'load' });
     const config = await page.evaluate(() => window.hlx.sidekick.config);
@@ -361,6 +168,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Uses outerHost from config', async () => {
+    const page = getPage();
     const testOuterHost = 'test-cdn-theblog.example.test';
     await mockStandardResponses(page, {
       configJs: `
@@ -378,6 +186,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Uses outerHost with branch in hlx3 mode', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/config-hlx3.html`, { waitUntil: 'load' });
     const outerHost = await page.evaluate(() => window.hlx.sidekick.config.outerHost);
@@ -385,6 +194,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Fails gracefully when unable to fetch status', async () => {
+    const page = getPage();
     await testPageRequests({
       page,
       url: `${fixturesPrefix}/config-default.html`,
@@ -406,6 +216,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Adds plugins via API', async () => {
+    const page = getPage();
     await page.goto(`${fixturesPrefix}/config-none.html`, { waitUntil: 'load' });
     await page.evaluate(() => {
       window.hlx.sidekick.add({
@@ -431,6 +242,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Loads config and plugins from project config', async () => {
+    const page = getPage();
     await mockStandardResponses(
       page, {
         configJs: `
@@ -451,6 +263,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Loads config and plugins from development environment', async () => {
+    const page = getPage();
     await mockStandardResponses(page, {
       configJs: `
         window.hlx.initSidekick({
@@ -469,6 +282,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Adds plugins from project (compatibility mode)', async () => {
+    const page = getPage();
     await mockStandardResponses(page, {
       pluginsJs: `
         window.hlx.sidekick.add({
@@ -485,6 +299,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Replaces plugin', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/config-default.html`, { waitUntil: 'load' });
     await page.evaluate(() => {
@@ -505,6 +320,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Extends plugin', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/config-default.html`, { waitUntil: 'load' });
     await page.evaluate(() => {
@@ -523,6 +339,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Removes plugin', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/config-default.html`, { waitUntil: 'load' });
     await page.evaluate(() => window.hlx.sidekick.remove('edit'));
@@ -531,6 +348,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Adds HTML element in plugin', async () => {
+    const page = getPage();
     await page.goto(`${fixturesPrefix}/config-none.html`, { waitUntil: 'load' });
     const text = await page.evaluate(() => {
       window.hlx.sidekick.add({
@@ -548,17 +366,19 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Loads custom CSS', async () => {
+    const page = getPage();
     await page.goto(`${fixturesPrefix}/config-none.html`, { waitUntil: 'load' });
     await page.evaluate(() => {
       window.hlx.sidekick.loadCSS('custom.css');
     });
     const bgColor = await page.$eval('div.hlx-sk',
       (elem) => window.getComputedStyle(elem).getPropertyValue('background-color'));
-    await sleep(3000);
+    await sleep();
     assert.strictEqual(bgColor, 'rgb(255, 255, 0)', 'Did not load custom CSS');
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Shows and hides notifications', async () => {
+    const page = getPage();
     await page.goto(`${fixturesPrefix}/config-none.html`, { waitUntil: 'load' });
 
     // shows notification
@@ -600,6 +420,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Close button hides sidekick', async () => {
+    const page = getPage();
     await page.goto(`${fixturesPrefix}/config-none.html`, { waitUntil: 'load' });
     await waitForEvent(page, 'hidden');
     await clickButton(page, 'close');
@@ -611,6 +432,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Share button copies sharing URL to clipboard', async () => {
+    const page = getPage();
     await page.goto(`${fixturesPrefix}/config-none.html`, { waitUntil: 'load' });
     await clickButton(page, 'share');
     assert.strictEqual(
@@ -620,437 +442,8 @@ describe('Test sidekick bookmarklet', () => {
     );
   }).timeout(IT_DEFAULT_TIMEOUT);
 
-  it('Edit plugin switches to editor from preview URL', async () => {
-    const apiMock = apiMocks.blog;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/edit-staging.html`,
-      check: (req) => {
-        if (req.url().startsWith('https://adobe.sharepoint.com/')) {
-          // check request to edit url
-          assert.ok(req.url() === apiMock.edit.url, 'Edit URL not called');
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'edit',
-      events: [
-        'statusfetched',
-        'contextloaded',
-        'envswitched',
-      ],
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Edit plugin switches to editor from production URL', async () => {
-    const apiMock = apiMocks.blog;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/edit-production.html`,
-      check: (req) => {
-        if (req.url().startsWith('https://adobe.sharepoint.com/')) {
-          // check request to edit url
-          assert.ok(req.url() === apiMock.edit.url, 'Edit URL not called');
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'edit',
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Preview plugin switches to preview from gdrive URL', async () => {
-    const apiMock = apiMocks.pages;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/preview-gdrive.html`,
-      check: (req) => {
-        if (req.url().includes('.hlx.page/')) {
-          // check request to preview url
-          assert.ok(
-            req.url() === `https://master--pages--adobe.hlx.page${apiMock.webPath}`,
-            'Preview URL not called',
-          );
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'preview',
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Preview plugin switches to preview from onedrive URL', async () => {
-    const apiMock = apiMocks.blog;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/preview-onedrive.html`,
-      check: (req) => {
-        if (req.url().includes('.hlx.page/')) {
-          // check request to preview url
-          assert.ok(
-            req.url() === `https://master--theblog--adobe.hlx.page${apiMock.webPath}`,
-            'Preview URL not called',
-          );
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'preview',
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Preview plugin updates preview when switching from editor', async () => {
-    const apiMock = apiMocks.blog;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/preview-onedrive-hlx3.html`,
-      check: (req) => {
-        if (req.method() === 'POST') {
-          // check post request to preview url
-          assert.ok(
-            req.url() === `https://admin.hlx3.page/preview/adobe/theblog/master${apiMock.webPath}`,
-            'Preview URL not updated',
-          );
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'preview',
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Preview plugin switches to preview from production URL', async () => {
-    const apiMock = apiMocks.blog;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/edit-production.html`,
-      check: (req) => {
-        if (req.url().includes('.hlx.page/')) {
-          // check request to preview url
-          assert.ok(
-            req.url() === `https://master--theblog--adobe.hlx.page${apiMock.webPath}`,
-            'Preview URL not called',
-          );
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'preview',
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Live plugin switches to live from gdrive URL', async () => {
-    const apiMock = apiMocks.pages;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/preview-gdrive.html`,
-      check: (req) => {
-        if (req.url().includes('.hlx.live/')) {
-          // check request to live url
-          assert.ok(
-            req.url() === `https://pages--adobe.hlx.live${apiMock.webPath}`,
-            'Live URL not called',
-          );
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'live',
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Production plugin switches to production from gdrive URL', async () => {
-    const apiMock = apiMocks.pages;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/preview-gdrive.html`,
-      check: (req) => {
-        if (req.url().includes('.adobe.com/')) {
-          // check request to production url
-          assert.ok(
-            req.url() === `https://pages.adobe.com${apiMock.webPath}`,
-            'Production URL not called',
-          );
-          return true;
-        }
-        // ignore otherwise
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'prod',
-    });
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Reload plugin sends purge request from preview URL and reloads page', async () => {
-    const apiMock = apiMocks.blog;
-    let loads = 0;
-    let purged = false;
-    let reloaded = false;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/reload-staging.html`,
-      check: (req) => {
-        if (!purged && req.method() === 'POST') {
-          // intercept purge request
-          const headers = req.headers();
-          purged = req.url() === `https://theblog--adobe.hlx.page${apiMock.webPath}`
-            && headers['x-forwarded-host'] === 'master--theblog--adobe.hlx.page';
-        } else if (req.url().endsWith('reload-staging.html')) {
-          loads += 1;
-          if (loads === 2) {
-            reloaded = true;
-            return true;
-          }
-        }
-        return false;
-      },
-      checkCondition: (request) => request.url().startsWith('https://')
-        || request.url().endsWith('reload-staging.html'),
-      mockResponses: [
-        apiMock,
-        purgeMock,
-      ],
-      plugin: 'reload',
-    });
-    // check result
-    assert.ok(purged, 'Purge request not sent');
-    assert.ok(reloaded, 'Reload not triggered');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Reload plugin uses preview API in hlx3 mode', async () => {
-    const apiMock = apiMocks.blog;
-    let loads = 0;
-    let apiCalled = false;
-    let reloaded = false;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/reload-staging-hlx3.html`,
-      check: (req) => {
-        if (!apiCalled && req.method() === 'POST') {
-          // intercept api request
-          apiCalled = req.url().endsWith(`/preview/adobe/theblog/master${apiMock.webPath}`);
-        } else if (req.url().endsWith('reload-staging-hlx3.html')) {
-          loads += 1;
-          if (loads === 2) {
-            // reload triggered
-            reloaded = true;
-            return true;
-          }
-        }
-        return false;
-      },
-      checkCondition: (request) => request.url().startsWith('https://')
-        || request.url().endsWith('reload-staging-hlx3.html'),
-      mockResponses: [
-        apiMock,
-        apiMock,
-      ],
-      plugin: 'reload',
-    });
-    // check result
-    assert.ok(apiCalled, 'Preview API not called');
-    assert.ok(reloaded, 'Reload not triggered');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Publish plugin sends purge request from preview URL and redirects to production URL', async () => {
-    const apiMock = apiMocks.blog;
-    let purged = false;
-    let redirected = false;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/publish-staging.html`,
-      check: (req) => {
-        if (!purged && req.method() === 'POST') {
-          // intercept purge request
-          const headers = req.headers();
-          purged = req.url() === `https://theblog--adobe.hlx.page${apiMock.webPath}`
-            && headers['x-forwarded-host'].split(',').length === 3;
-        } else if (req.url().startsWith('https://blog.adobe.com')) {
-          // intercept redirect to production
-          redirected = true;
-          return true;
-        }
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-        purgeMock,
-      ],
-      plugin: 'publish',
-    });
-    // check result
-    assert.ok(purged, 'Purge request not sent');
-    assert.ok(redirected, 'Redirect not sent');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Publish plugin also purges dependencies', async () => {
-    const apiMock = apiMocks.blog;
-    const dependencies = [
-      '/en/topics/foo.html',
-      'bar.html?step=1',
-    ];
-    const purged = [];
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/publish-staging.html`,
-      prep: async (p) => {
-        // add publish dependencies
-        await p.evaluate((deps) => {
-          window.hlx.dependencies = deps;
-        }, dependencies);
-      },
-      check: (req) => {
-        if (req.method() === 'POST') {
-          // check result
-          const purgeUrl = new URL(req.url());
-          purged.push(`${purgeUrl.pathname}${purgeUrl.search}`);
-        }
-        return purged.length === 3;
-      },
-      mockResponses: [
-        apiMock,
-        purgeMock,
-        purgeMock,
-        purgeMock,
-      ],
-      plugin: 'publish',
-    });
-    assert.deepStrictEqual(purged, [
-      '/en/topics/bla.html',
-      '/en/topics/foo.html',
-      '/en/topics/bar.html?step=1',
-    ], 'Purge request not sent');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Publish plugin refuses to publish without production host', async () => {
-    const apiMock = apiMocks.blog;
-    let noPurge = true;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/publish-staging.html`,
-      prep: async (p) => {
-        // remove production host from config
-        await p.evaluate(() => {
-          delete window.hlx.sidekick.config.outerHost;
-          delete window.hlx.sidekick.config.host;
-        });
-      },
-      check: (req) => {
-        if (req.method() === 'POST') {
-          // intercept purge request
-          noPurge = false;
-          return true;
-        }
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-      ],
-      plugin: 'publish',
-      timeout: 5000,
-    });
-    assert.ok(noPurge, 'Did not purge inner host only');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Publish plugin uses publish API in hlx3 mode', async () => {
-    const apiMock = apiMocks.blog;
-    let apiCalled = false;
-    let redirected = false;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/publish-staging-hlx3.html`,
-      check: (req) => {
-        if (!apiCalled && req.method() === 'POST') {
-          // intercept purge request
-          apiCalled = req.url().endsWith('/live/adobe/theblog/main/en/topics/bla.html');
-        } else if (req.url().startsWith('https://blog.adobe.com')) {
-          redirected = true;
-          return true;
-        }
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-        apiMock,
-        htmlMock,
-        htmlMock,
-      ],
-      plugin: 'publish',
-    });
-    // check result
-    assert.ok(apiCalled, 'Purge request not sent');
-    assert.ok(redirected, 'Redirect not sent');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('Publish plugin redirects to live instead of bring-your-own-CDN production host', async () => {
-    const apiMock = apiMocks.blog;
-    let redirected = false;
-    await testPageRequests({
-      page,
-      url: `${fixturesPrefix}/publish-staging.html`,
-      prep: async (p) => {
-        // set byocdn flag
-        await p.evaluate(() => {
-          window.hlx.sidekick.config.byocdn = true;
-        });
-      },
-      check: (req) => {
-        if (req.url().startsWith('https://theblog--adobe.hlx.live')) {
-          // intercept redirect to live
-          redirected = true;
-          return true;
-        }
-        return false;
-      },
-      mockResponses: [
-        apiMock,
-        purgeMock,
-      ],
-      plugin: 'publish',
-    });
-    // check result
-    assert.ok(redirected, 'Redirect to live not sent');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
-  it('No publish plugin on bring-your-own-CDN production host', async () => {
-    await mockStandardResponses(page);
-    // open test page
-    await page.goto(`${fixturesPrefix}/publish-byocdn.html`, { waitUntil: 'load' });
-    const plugins = await getPlugins(page);
-    assert.ok(!plugins.includes('publish'), 'Unexpected publish plugin found');
-  }).timeout(IT_DEFAULT_TIMEOUT);
-
   it('Development environment is correctly detected', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/is-dev.html`, { waitUntil: 'load' });
     assert.ok(
@@ -1060,8 +453,9 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Preview environment is correctly detected', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
-    await page.goto(`${fixturesPrefix}/publish-staging.html`, { waitUntil: 'load' });
+    await page.goto(`${fixturesPrefix}/is-preview.html`, { waitUntil: 'load' });
     assert.ok(
       await page.evaluate(() => window.hlx.sidekick.isInner() && window.hlx.sidekick.isHelix()),
       'Did not detect preview URL',
@@ -1077,6 +471,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Live environment is correctly detected', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/is-live.html`, { waitUntil: 'load' });
     assert.ok(
@@ -1094,6 +489,7 @@ describe('Test sidekick bookmarklet', () => {
   }).timeout(IT_DEFAULT_TIMEOUT);
 
   it('Production environment is correctly detected', async () => {
+    const page = getPage();
     await mockStandardResponses(page);
     await page.goto(`${fixturesPrefix}/edit-production.html`, { waitUntil: 'load' });
     assert.ok(
