+++ test/fixtures/is-preview.html
@@ -1,0 +1,14 @@
+<html>
+  <head>
+    <meta charset="utf-8"/>
+  </head>
+  <body>
+    <!-- switch to production from staging URL -->
+    <input type="hidden" id="sidekick_test_location" value="https://theblog--adobe.hlx.page/en/topics/bla.html"/>
+    <script
+      id='hlx-sk-test'
+      src='./bookmarklet.js'
+      data-config='{"owner":"adobe","repo":"theblog","ref":"master","host":"blog.adobe.com"}'
+    ></script>
+  </body>
+</html>
